Address a bunch of review comments
