+++ CHANGES.md
@@ -4,7 +4,7 @@
 
 ##### Additions :tada:
 
-- Added new TilesetContentLoaders constructible using the new `TilesetLoaderFactory` constructor on `Cesium3DTilesSelection::Tileset`.
+- Added new TilesetContentLoaders constructible using the new `TilesetContentLoaderFactory` constructor on `Cesium3DTilesSelection::Tileset`.
   - `ITwinCesiumCuratedContentLoader` can load tilesets from the [iTwin Cesium Curated Content API](https://developer.bentley.com/apis/cesium-curated-content/).
   - `IModelMeshExportContentLoader` can load [iModels](https://www.itwinjs.org/learning/imodels/) exported to the 3D Tiles format through the [Mesh Export API](https://developer.bentley.com/apis/mesh-export/).
   - `ITwinRealityDataContentLoader` can load 3D Tiles iTwin Reality Data through the [Reality Management API](https://developer.bentley.com/apis/reality-management/overview/).
