+++ CesiumITwinClient/src/IModel.cpp
@@ -1,0 +1,13 @@
+#include <CesiumITwinClient/IModel.h>
+
+namespace CesiumITwinClient {
+IModelState iModelStateFromString(const std::string& str) {
+  if (str == "initialized") {
+    return IModelState::Initialized;
+  } else if (str == "notInitialized") {
+    return IModelState::NotInitialized;
+  }
+
+  return IModelState::Unknown;
+}
+} // namespace CesiumITwinClient
