+++ doc/CMakeLists.txt
@@ -53,7 +53,7 @@ if(DOXYGEN_FOUND)
         "CESIUM3DTILESSELECTION_API"
         "CESIUM3DTILESWRITER_API"
         "CESIUMASYNC_API"
-        "CESIUMCLIENTCOMMOn_API"
+        "CESIUMCLIENTCOMMON_API"
         "CESIUMGEOMETRY_API"
         "CESIUMGEOSPATIAL_API"
         "CESIUMGLTF_API"
